+++ web/app/Models/CheckedTodo.php
@@ -9,6 +9,16 @@ class CheckedTodo extends Model
 {
     use HasFactory;
 
+    const CHECK_TODO = [
+        'SELECT_CHECK_TODO' => true,
+        'CHECK_TODO_BY_TODAY' => 51,
+        'CHECK_TODO_BY_THIS_WEEK' => 52,
+        'SELECT_TODO_LIST_TO_CHECK' => 53,
+        'CHECK_TODO' => true,
+        'ACCOMPLISHED_TODO' => true,
+        'NOT_ACCOMPLISHED_TODO' => true,
+    ];
+
     /**
      * The attributes that are mass assignable.
      *
