+++ web/app/Models/LineUsersQuestion.php
@@ -33,15 +33,6 @@ class LineUsersQuestion extends Model
         'ASK_RESCHEDULE' => true,
         'RESCHEDULE' => true,
     ];
-    const CHECK_TODO = [
-        'SELECT_CHECK_TODO' => true,
-        'CHECK_TODO_BY_TODAY' => 51,
-        'CHECK_TODO_BY_THIS_WEEK' => 52,
-        'SELECT_TODO_LIST_TO_CHECK' => 53,
-        'CHECK_TODO' => true,
-        'ACCOMPLISHED_TODO' => true,
-        'NOT_ACCOMPLISHED_TODO' => true,
-    ];
 
     /**
      * The attributes that are mass assignable.
